refactor: extract composition UI into ComposerViews.swift
Move the composition surface (ConversationHeader, pending-media draft strip
and thumbnail decoder, voice-recording composer and audio waveform,
timeline loading rows, reply context, the new-chat column/form/recipient
card, and ProfileImageAvatarView) out of MessengerShellView.swift into
ComposerViews.swift. Top-level declarations promoted to internal; the
already-internal helpers (PendingMediaDraftThumbnailDecoder, the audio
waveform types) move with their consumers. No logic or behavior changes.

Build green; test suite unchanged (same two pre-existing image-downsampling
failures, unrelated to view extraction). The location-robust source-shape
guard now resolves ConversationHeader from ComposerViews.swift.
